Move torch variations into tile data

+++ src/client/java/minicraft/level/tile/DoorTile.java
@@ -52,7 +52,7 @@ public class DoorTile extends Tile {
 			if (tool.type == type.getRequiredTool()) {
 				if (player.payStamina(4 - tool.level) && tool.payDurability()) {
 					int data = level.getData(xt, yt);
-					level.setTile(xt, yt, Tiles.get(id + 3)); // Will get the corresponding floor tile.
+					level.setTile(xt, yt, Tiles.get((short) (id + 3))); // Will get the corresponding floor tile.
 					Sound.play("monsterhurt");
 					level.dropItem(xt * 16 + 8, yt * 16 + 8, Items.get(type.name() + " Door"));
 					AdvancementElement.AdvancementTrigger.ItemUsedOnTileTrigger.INSTANCE.trigger(
